- make upgrade script more robust to allow multiple runs

+++ packages/acs-kernel/sql/postgresql/upgrade/upgrade-5.9.1d3-5.9.1d4.sql
@@ -6,8 +6,11 @@ BEGIN
    select setting from pg_settings where name='server_version_num' and setting::int >= 90200 into v_dummy;
    IF found THEN
 
-      	drop view IF EXISTS anon_func_seq;
-   	ALTER SEQUENCE IF EXISTS t_anon_func_seq RENAME TO anon_func_seq;
+        select 1 from pg_views where viewname = 'anon_func_seq' into v_dummy;
+	IF v_dummy THEN
+      	   drop view IF EXISTS anon_func_seq;
+   	   ALTER SEQUENCE IF EXISTS t_anon_func_seq RENAME TO anon_func_seq;
+	END IF;
 
    ELSE
 	-- verison earlier than 9.2, no "IF EXISTS"
